3_real_life_application/lsc-2-emotions-detection/requirements.txt
parliament-functions==0.1.0
boto3
fer
pandas
pillow>=7.0.0 
numpy>=1.18.1 
opencv-python>=4.1.2